Fix AG.5 merge-forward conflicts and CI drift
