+++ crates/atm-daemon/src/peer_transport/mod.rs
@@ -994,22 +994,9 @@ impl PeerTransportRuntime {
     ) -> Result<(), AtmError> {
         self.server.start(dispatcher).map(|_| ())
     }
-
     pub(crate) fn shutdown(&self) -> Result<(), AtmError> {
         self.server.shutdown()
     }
-
-    #[allow(dead_code, reason = "retained for existing peer-transport tests")]
-    pub(crate) fn reload_listener(
-        &self,
-        listen_addr: Option<SocketAddr>,
-        dispatcher: Arc<dyn RequestDispatcher + Send + Sync>,
-    ) -> Result<(), AtmError> {
-        self.server
-            .reload(listen_addr.into_iter().collect::<Vec<_>>(), dispatcher)?;
-        Ok(())
-    }
-
     pub(crate) fn reload_listeners(
         &self,
         listen_addrs: Vec<SocketAddr>,
@@ -1017,11 +1004,9 @@ impl PeerTransportRuntime {
     ) -> Result<Vec<server::PeerListenerOutcome>, AtmError> {
         self.server.reload(listen_addrs, dispatcher)
     }
-
     pub(crate) fn bound_addr(&self) -> Result<Option<SocketAddr>, AtmError> {
-        self.server.bound_addr()
+        server::bound_addr(&self.server)
     }
-
     #[cfg(test)]
     pub(crate) fn new_for_test(
         endpoint: SocketAddr,
